test(plugin): reorganized test fixtures
