--- jest.config.js
@@ -1,11 +1,0 @@
-'use strict';
-
-module.exports = {
-  roots: [
-  	'<rootDir>/test/',
-  	'<rootDir>/src/'
-  ],
-  collectCoverageFrom: [
-    'src/**/*.js'
-  ]
-}
