+++ src/__tests__/api.js
@@ -1,5 +1,5 @@
 import { withPrefix } from 'gatsby'
-import routes from '../../test/__fixtures__/routes'
+import routes from './__fixtures__/routes'
 import {
   getRoutes,
   routeExists,
