+++ src/components/Link/__tests__/Link.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import TestRenderer from 'react-test-renderer'
 import Link from '../Link'
-import routes from '../../../../test/__fixtures__/routes'
+import routes from '../../../__tests__/__fixtures__/routes'
 
 // Create a virtual mock for routes.js
 jest.doMock('../../../routes', () => routes, { virtual: true })
