+++ src/gatsby/__tests__/__fixtures__/pagesSchema.js
