+++ src/gatsby/__tests__/__fixtures__/pluginOptionsSchema.js
