+++ src/gatsby/__tests__/__fixtures__/source-nodes.js
