+++ src/gatsby/__tests__/create-pages.js
@@ -2,7 +2,7 @@ import fs from 'fs'
 import path from 'path'
 import { createPages } from '../'
 import { mountOptions, mountFile, mountDir, mountModule } from '../../../test/helpers'
-import testCases from '../../../test/__fixtures__/create-pages'
+import testCases from './__fixtures__/create-pages'
 
 // Use in-memory file system
 jest.mock('fs')
