+++ src/gatsby/__tests__/pluginOptionsSchema.js
@@ -1,6 +1,6 @@
 import { testPluginOptionsSchema } from 'gatsby-plugin-utils'
 import { pluginOptionsSchema } from '../'
-import testCases from '../../../test/__fixtures__/pluginOptionsSchema'
+import testCases from './__fixtures__/pluginOptionsSchema'
 
 describe('pluginOptionsSchema', () => {
   for (const { title, options } of testCases) {
